Create Computations with User Input.asm
This is assembly code that will prompt the user for a 3 digit number, and after receiving that number, the program will then use that number for different computations like finding out what the squared value of that number would be, the cubed value, and also checking to see if that number is a Narcissistic number. The program also verifies that the number is 3 digits long.