Random Number Generator.asm
This is assembly code for a random number generator. It starts from a seed number and will generate a number based off of that number. But in order to make it random, we us TSC (Time Stamp Counter) to give us a random seed number to work with. While just by using the TSC value in itself seems "random", its just a 64-bit register time stamp of the number of clock cycles since the last reset. So its not truly random.

Below line 70 is a use for this random number generator, like using it to make a random string of letters. This could be used for encryption purposes.